==> 02_create_balanced_sample_translation.ipynb <==
# Balanced Sampling of News Articles by Country

This script performs the following steps to prepare a balanced news dataset:

- **Load and prepare data:**  
  Load news articles from a specified folder, filtering by selected countries (`Bulgaria`, `Italy`, `Netherlands`, `United Kingdom`) and relevant news outlets using the `load_and_prepare_data` function.

- **Filter results:**  
  Outputs the total number of articles remaining after filtering non-relevant outlets.

- **Create balanced sample:**  
  Generate a balanced sample of news articles (`total_samples = 2000`) evenly distributed across the selected countries using the `balanced_sample` function.

- **Summary and preview:**  
  Prints the size of the sample, counts of articles per country, and previews the first few rows of the sampled dataset.

- **Save to CSV:**  
  Saves the balanced sample as a CSV file for later use in analysis or modeling.


import os
from dataloader import load_and_prepare_data, balanced_sample, load_human_annotated_for_translation
from config import NEWS_FOLDER, SELECTED_COUNTRIES

# Create a balanced sample, e.g. total 50 samples across countries
data = load_and_prepare_data(
    news_folder="~/webdav/ASCOR-FMG-5580-RESPOND-news-data (Projectfolder)/", 
    countries=["Bulgaria", "Italy", "Netherlands", "United_Kingdom"], 
    outlet_dir="~/webdav/ASCOR-FMG-5580-RESPOND-news-data (Projectfolder)/selected_outlets/"
)

print(f'N of all news articles after filtering out non-relevant outlets: {len(data)}')
#data.to_csv('~/webdav/ASCOR-FMG-5580-RESPOND-news-data (Projectfolder)/output/news_filtered_Bulgaria_Italy_Netherlands_United_Kingdom_outlets.csv')

total_samples = 10000
df_news_sample = balanced_sample(data, total_samples=total_samples, countries=SELECTED_COUNTRIES)

print(f"Sample created with {len(df_news_sample)} rows")
print("Samples per country:")
print(df_news_sample["country"].value_counts())

print("\nSample preview:")
print(df_news_sample.head())

sample_csv_path = f"~/webdav/ASCOR-FMG-5580-RESPOND-news-data (Projectfolder)/output/news_sample_{total_samples}.csv"
df_news_sample.to_csv(sample_csv_path, index=False)
print(f"\nSample saved to: {sample_csv_path}")
